Don't strip out component stack in assertConsole helpers (#30204)
Use the same normalizeCodeLocInfo that we use everywhere else.

We should actually test the component stack itself. Not just that it
exists. This was causing false passes.

However, the logic was also wrong before because it wouldn't always
strip out the last line so wouldn't accurately normalize it. Leading to
false failures as well.