[DevTools] Support REACT_LEGACY_ELEMENT_TYPE for formatting JSX (#30779)
DevTools shouldn't use react-is since that's versioned to one version of
React. We don't need to since we use all the symbols from
shared/ReactSymbols anyway and have a fork of typeOf that can cover
both.

Now JSX of old React versions show up with proper JSX formatting when
inspecting.